cmake:   Fixed a bunch of warnings

+++ modules/tracktion_graph/tracktion_graph/tracktion_graph_Node.h
@@ -302,7 +302,7 @@ inline void Node::process (juce::Range<int64_t> referenceSampleRange)
 
     const int numSamples = (int) referenceSampleRange.getLength();
     jassert (numSamples > 0); // This must be a valid number of samples to process
-    jassert (numChannelsBeforeProcessing == 0 || numSamples <= audioBuffer.getNumFrames());
+    jassert (numChannelsBeforeProcessing == 0 || numSamples <= (int) audioBuffer.getNumFrames());
 
     auto inputBlock = numChannelsBeforeProcessing > 0 ? juce::dsp::AudioBlock<float> (toAudioBlock (audioBuffer)).getSubBlock (0, (size_t) numSamples)
                                                       : juce::dsp::AudioBlock<float>();
